refactoring, traces in mission generator, debug to separate file
